Bug 1588233 - P3: Update soundtouch to version 2.1.2. r=padenot
The changed-files are updated by running `sh update.sh <upstream-path>`
with moz-libsoundtouch.patch applied.

Differential Revision: https://phabricator.services.mozilla.com/D49029

UltraBlame original commit: 41cd531e5e837a1aee11aa7ae317c386dd5c166a

+++ media/libsoundtouch/src/InterpolateCubic.h
@@ -29,10 +29,6 @@
 
 
 
-
-
-
-
 #ifndef _InterpolateCubic_H_
 #define _InterpolateCubic_H_
 
@@ -56,7 +52,7 @@ protected:
                         const SAMPLETYPE *src, 
                         int &srcSamples);
 
-    float fract;
+    double fract;
 
 public:
     InterpolateCubic();
